Support parsing of multiline paragraphs
To model this, the commit renames the class Text to Paragraph.
In addition, change all Token filter methods to use blocks
instead of token kind parameters.

Currently, one test case fails expectedly. The case tests the number of
elements for a paragraph that contains bold and italic text, which is
currently not supported by the parser.