Add a simple preview mode
Rename the main script to silas to reflect the change that it does more
than simply building the website and add a preview mode that sets the
url base to the path of the generated folder.
This is a total hackjob and I want to rewrite the URL system to generate
relative urls. But this is easier for now.