Exclude drafts from build
Move draft property to OrgObject. If it resolves to true the file will
only be included in preview mode. When a link points to a draft a
warning will be printed and the link will be only printed as text in
build mode.

Move the flag for the preview mode to a Singleton class to be available
from everywhere.