Reintroduce rendering of header
The header now contains links to all pages and projects that are no
drafts. Also introduce a Preamble property `hidden` for projects. If it
is set in the index file, the project won't be linked to in the header.
To change the text of the link the property alias can be set.

Resolve the link to the index page manually and fix the css path for the
website's index page.