Defer HTML export of orgfile to generation stage
Previously an org file was parsed as html directly when loading. To
properly resolve internal links (links to other org files), all files
have to be loaded before the HTML code can be rendered. To do this,
store the original content of the org file and add a function `to_html`
that renders the file to HTML.