Show posts ordered by date in Blog index
Use chrono::naive::NaiveDate instead of String to store dates in Post
structs and implement the Ord trait (and dependencies) for
SerializedPost based on published. If published is None, it is
interpreted to be greater than each date. This has the effect, that all
posts without a publishing date (which are interpreted as drafts) are
shown before posts that have a publishing date.

Since the date isn't stored as a formatted string anymore, add
formatting in the `date` helper for handlebars.