Fix panic in preview mode when output folder doesn't exist
The PreviewBuilder uses fs::canonicalize to get the absolute path of the
output directory. If it doesn't exist, it throws an error, which results
to a panic in the current codebase.
To prevent panic make folder creation a static method and call it before
instantiating builder. This creates a folder even when generation for a
single file is called. Leave this unfixed for now.