Implement rendering of latex
Create a script that clones the most recent version of mathjax and add it
to the theme directory. To make mathjax available in the generated
output, implement a recursive directory copy function.

I implemented the parsing of latex expressions in the fork of orgize.
Point the dependency in Cargo.toml to my fork.