Remove ignored files and paths from website
When loading the website files are no longer ignored. This also means
that now the website index is rendered (previously, the about page was
mirrored as the website index). While testing with my website, the
rendering failed due to an invalid image path. I improved the error
message for this case.