Create temporary directory during website generation
Using a temporary directory makes it possible to only override the
generated result in case the full generation was successful.

Currently, the website generation uses assert and panic all over the
place. In order to catch the errors at runtime (which is necessary for
a gentle exit after a failed generation), reorganise all their
occurences into error types.